chore : split ScheduleForm modal
